+++ apps/example/pages/pagerouter/admin/[[...nextadmin]].tsx
@@ -1,14 +1,15 @@
 import { AdminComponentProps, NextAdmin } from "@premieroctet/next-admin";
-import "@premieroctet/next-admin/dist/styles.css";
 import { GetServerSideProps, GetServerSidePropsResult } from "next";
 import Dashboard from "../../../components/Dashboard";
 import { options } from "../../../pageRouterOptions";
 import { prisma } from "../../../prisma";
 import schema from "../../../prisma/json-schema/json-schema.json";
+import "../../../styles.css";
 
 const pageOptions = options;
 
 export default function Admin(props: AdminComponentProps) {
+  
   return <NextAdmin {...props} dashboard={Dashboard} options={pageOptions} />;
 }