+++ packages/next-admin/src/components/Form.tsx
@@ -5,16 +5,19 @@ import {
   BaseInputTemplateProps,
   ErrorSchema,
   FieldTemplateProps,
+  ObjectFieldTemplateProps,
   SubmitButtonProps,
-  getSubmitButtonOptions,
+  getSubmitButtonOptions
 } from "@rjsf/utils";
 import validator from "@rjsf/validator-ajv8";
 import clsx from "clsx";
-import { ChangeEvent, cloneElement, useMemo, useState } from "react";
+import React, { ChangeEvent, cloneElement, useMemo, useState } from "react";
 import { useConfig } from "../context/ConfigContext";
+import { FormContext, FormProvider } from "../context/FormContext";
+import { useI18n } from "../context/I18nContext";
 import { PropertyValidationError } from "../exceptions/ValidationError";
 import { useRouterInternal } from "../hooks/useRouterInternal";
-import { Field, ModelAction, ModelName, SubmitFormResult } from "../types";
+import { Field, ModelAction, ModelName, NextAdminOptions, SubmitFormResult } from "../types";
 import { getSchemas } from "../utils/jsonSchema";
 import ActionsDropdown from "./ActionsDropdown";
 import ArrayField from "./inputs/ArrayField";
@@ -22,9 +25,10 @@ import CheckboxWidget from "./inputs/CheckboxWidget";
 import DateTimeWidget from "./inputs/DateTimeWidget";
 import DateWidget from "./inputs/DateWidget";
 import FileWidget from "./inputs/FileWidget";
-import SelectWidget from "./inputs/SelectWidget";
-import Button from "./radix/Button";
 import JsonField from "./inputs/JsonField";
+import SelectWidget from "./inputs/SelectWidget";
+import TextareaWidget from "./inputs/TextareaWidget";
+import Button from "./radix/Button";
 
 // Override Form functions to not prevent the submit
 class CustomForm extends RjsfForm {
@@ -47,6 +51,7 @@ export type FormProps = {
   validation?: PropertyValidationError[];
   action?: (formData: FormData) => Promise<SubmitFormResult | undefined>;
   title: string;
+  options?: Required<NextAdminOptions>["model"][ModelName];
   customInputs?: Record<Field<ModelName>, React.ReactElement | undefined>;
   actions?: ModelAction[];
 };
@@ -61,6 +66,7 @@ const widgets: CustomForm["props"]["widgets"] = {
   SelectWidget: SelectWidget,
   CheckboxWidget: CheckboxWidget,
   FileWidget: FileWidget,
+  TextareaWidget: TextareaWidget,
 };
 
 const Form = ({
@@ -70,6 +76,7 @@ const Form = ({
   resource,
   validation: validationProp,
   action,
+  options,
   title,
   customInputs,
   actions,
@@ -78,13 +85,10 @@ const Form = ({
   const { edit, id, ...schemas } = getSchemas(data, schema, dmmfSchema);
   const { basePath } = useConfig();
   const { router } = useRouterInternal();
+  const { t } = useI18n();
   const submitButton = (props: SubmitButtonProps) => {
     const { uiSchema } = props;
-    const {
-      norender,
-      submitText,
-      props: buttonProps,
-    } = getSubmitButtonOptions(uiSchema);
+    const { norender, props: buttonProps } = getSubmitButtonOptions(uiSchema);
     if (norender) {
       return null;
     }
@@ -92,19 +96,19 @@ const Form = ({
     return (
       <div className="flex space-x-2 mt-4 justify-between">
         <div className="flex space-x-2">
-          <Button {...buttonProps}>{submitText}</Button>
-          {edit && (
-            <Button
-              type="submit"
-              name="__admin_action"
-              value="delete"
-              variant="destructive"
-            >
-              Delete
-            </Button>
-          )}
+          <Button {...buttonProps} name='__admin_redirect' value="list">{t("form.button.save.label")}</Button>
+          <Button {...buttonProps} variant={'ghost'}>{t("form.button.save_edit.label")}</Button>
         </div>
-
+        {edit && (
+          <Button
+            type="submit"
+            name="__admin_action"
+            value="delete"
+            variant="destructive"
+          >
+            {t("form.button.delete.label")}
+          </Button>
+        )}
       </div>
     );
   };
@@ -140,10 +144,10 @@ const Form = ({
           },
         });
       }
-
       if (result?.created) {
+        const pathname = result?.redirect ? `${basePath}/${resource.toLowerCase()}` : `${basePath}/${resource.toLowerCase()}/${result.createdId}`
         return router.replace({
-          pathname: `${basePath}/${resource.toLowerCase()}/${result.createdId}`,
+          pathname,
           query: {
             message: JSON.stringify({
               type: "success",
@@ -154,12 +158,16 @@ const Form = ({
       }
 
       if (result?.updated) {
-        location.search = `?message=${encodeURIComponent(
-          JSON.stringify({
-            type: "success",
-            content: "Updated successfully",
-          })
-        )}`;
+        const pathname = result?.redirect ? `${basePath}/${resource.toLowerCase()}` : location.pathname
+        return router.replace({
+          pathname,
+          query: {
+            message: JSON.stringify({
+              type: "success",
+              content: "Updated successfully",
+            }),
+          },
+        });
       }
 
       if (result?.error) {
@@ -187,19 +195,33 @@ const Form = ({
           errors,
           children,
         } = props;
+        const labelAlias = options?.aliases?.[id as Field<typeof resource>] || label;
+        let styleField = options?.edit?.styles?.[id as Field<typeof resource>]
+        const sanitizedClassNames = classNames
+          ?.split(",")
+          .filter((className) => className !== "null")
+          .join(" ");
         return (
-          <div className={clsx(classNames, "py-1")} style={style}>
+          <div style={style} className={clsx(sanitizedClassNames, styleField)}>
             <label
-              className="block text-sm font-medium leading-6 text-gray-900 capitalize"
+              className={clsx("block text-sm font-medium leading-6 text-gray-900 capitalize")}
               htmlFor={id}
             >
-              {label}
+              {labelAlias}
               {required ? "*" : null}
             </label>
             {description}
             {children}
             {errors}
             {help}
+          </div>
+        );
+      },
+      ObjectFieldTemplate: (props: ObjectFieldTemplateProps) => {
+        const styleForm = options?.edit?.styles?._form
+        return (
+          <div className={clsx('grid', styleForm)}>
+            {props.properties.map((element) => element.content)}
           </div>
         );
       },
@@ -246,6 +268,7 @@ const Form = ({
           <input
             onChange={onChangeOverride || onTextChange}
             {...props}
+            value={props.value ?? ""}
             className={clsx(
               "block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 px-2 disabled:opacity-50 disabled:bg-gray-200 disabled:cursor-not-allowed",
               { "ring-red-600": rawErrors }
@@ -258,7 +281,15 @@ const Form = ({
       },
       FieldErrorTemplate: ({ errors }) => {
         return errors ? (
-          <div className="text-sm text-red-600 mt-1">{errors}</div>
+          <div className="text-sm text-red-600 mt-1">
+            {errors.map((error, idx) => {
+              if (typeof error === "string") {
+                return <React.Fragment key={idx}>{t(error)}</React.Fragment>;
+              }
+
+              return <React.Fragment key={idx}>{error}</React.Fragment>;
+            })}
+          </div>
         ) : null;
       },
     }),
@@ -280,26 +311,32 @@ const Form = ({
           />
         )}
       </div>
-      <CustomForm
-        // @ts-expect-error
-        action={action ? onSubmit : ""}
-        method="post"
-        idPrefix=""
-        idSeparator=""
-        enctype={!action ? "multipart/form-data" : undefined}
-        {...schemas}
-        formData={data}
-        validator={validator}
-        extraErrors={extraErrors}
-        fields={fields}
-        templates={{
-          ...templates,
-          ButtonTemplates: { SubmitButton: submitButton },
-        }}
-        widgets={widgets}
-        onSubmit={(e) => console.log("onSubmit", e)}
-        onError={(e) => console.log("onError", e)}
-      />
+      <FormProvider initialValue={data}>
+        <FormContext.Consumer>
+          {({ formData, setFormData }) =>
+            <CustomForm
+              // @ts-expect-error
+              action={action ? onSubmit : ""}
+              {...(!action ? { method: "post" } : {})}
+              onChange={(e) => { setFormData(e.formData) }}
+              idPrefix=""
+              idSeparator=""
+              enctype={!action ? "multipart/form-data" : undefined}
+              {...schemas}
+              formData={formData}
+              validator={validator}
+              extraErrors={extraErrors}
+              fields={fields}
+              templates={{
+                ...templates,
+                ButtonTemplates: { SubmitButton: submitButton },
+              }}
+              widgets={widgets}
+              onSubmit={(e) => console.log("onSubmit", e)}
+              onError={(e) => console.log("onError", e)}
+            />}
+        </FormContext.Consumer>
+      </FormProvider>
     </div>
   );
 };