+++ packages/next-admin/src/utils/prisma.ts
@@ -24,23 +24,40 @@ export const createWherePredicate = (
 ) => {
   return search
     ? {
-        OR: fieldsFiltered
-          ?.filter((field) => field.kind === "scalar")
-          .map((field) => {
-            if (field.type === "String") {
-              return {
-                [field.name]: { contains: search, mode: "insensitive" },
-              };
-            }
-            if (field.type === "Int" && !isNaN(Number(search))) {
-              return { [field.name]: Number(search) };
-            }
-            return null;
-          })
-          .filter(Boolean),
-      }
+      OR: fieldsFiltered
+        ?.filter((field) => field.kind === "scalar")
+        .map((field) => {
+          if (field.type === "String") {
+            // @ts-ignore
+            const mode = Prisma?.QueryMode ? { mode: Prisma.QueryMode.insensitive } : {};
+            return {
+              [field.name]: { contains: search, ...mode },
+            };
+          }
+          if (field.type === "Int" && !isNaN(Number(search))) {
+            return { [field.name]: Number(search) };
+          }
+          return null;
+        })
+        .filter(Boolean),
+    }
     : {};
 };
+
+export const isSatisfyingSearch = (
+  item: any,
+  fieldsFiltered?: Prisma.DMMF.Field[],
+  search?: string,
+  withFormatter: boolean = false) => {
+  const fieldsFilteredNames = fieldsFiltered
+    ?.filter((field) => field.kind === "scalar")
+    ?.map((field) => field.name) ?? [];
+  withFormatter && fieldsFilteredNames?.push('_formatted')
+  if (!fieldsFilteredNames || !search) return true;
+  return fieldsFilteredNames?.reduce((acc, field) => {
+    return item[field].toString().toLowerCase().includes(search?.trim().toLowerCase()) || acc;
+  }, false);
+}
 
 export const preparePrismaListRequest = <M extends ModelName>(
   resource: M,
@@ -79,22 +96,7 @@ export const preparePrismaListRequest = <M extends ModelName>(
   let fieldsFiltered = model?.fields;
   const list = options?.model?.[resource]?.list as ListOptions<M>;
   if (list) {
-    const listDisplayedKeys = list.display;
-    select = listDisplayedKeys?.reduce(
-      (acc, column) => {
-        const field = model?.fields.find(({ name }) => name === column);
-        if (field?.kind === "object" && field?.isList === true) {
-          if (!acc._count) acc._count = { select: {} };
-          acc._count.select = { ...acc._count.select, [column]: true };
-        } else {
-          // @ts-expect-error
-          acc[column] = true;
-        }
-        return acc;
-      },
-      { [getModelIdProperty(resource)]: true } as Select<M>
-    );
-
+    select = selectPayloadForModel(resource, list, 'object');
     fieldsFiltered =
       model?.fields.filter(({ name }) =>
         list.search?.includes(name as Field<M>)
@@ -117,7 +119,7 @@ export const getMappedDataList = async (
   options: NextAdminOptions,
   searchParams: URLSearchParams,
   context: NextAdminContext,
-  appDir = false
+  appDir = false,
 ) => {
   const prismaListRequest = preparePrismaListRequest(
     resource,
@@ -156,11 +158,9 @@ export const getMappedDataList = async (
   data.forEach((item, index) => {
     Object.keys(item).forEach((key) => {
       let itemValue;
-
+      const model = capitalize(key) as ModelName;
+      const idProperty = getModelIdProperty(model);
       if (typeof item[key] === "object" && item[key] !== null) {
-        const model = capitalize(key) as ModelName;
-        const idProperty = getModelIdProperty(model);
-
         switch (item[key].type) {
           case "link":
             itemValue = item[key].value.label;
@@ -195,10 +195,18 @@ export const getMappedDataList = async (
         item[key].__nextadmin_formatted = listFields[
           key as keyof typeof listFields
           // @ts-expect-error
-        ]?.formatter?.(itemValue ?? item[key], context);
+        ]?.formatter?.(itemValue ?? item[key], context)
       } else {
+        if (typeof item[key]?.__nextadmin_formatted === 'object') {
+          item[key].__nextadmin_formatted = item[key].__nextadmin_formatted[idProperty]
+        }
         data[index][key] = item[key];
       }
+
+      if (typeof item[key]?.value === 'object') {
+        item[key].value.label = item[key].value.label[idProperty]
+      }
+
     });
   });
 
@@ -208,3 +216,26 @@ export const getMappedDataList = async (
     error,
   };
 };
+
+export const selectPayloadForModel = <M extends ModelName>(resource: M, options?: EditOptions<M> | ListOptions<M>, level: 'scalar' | 'object' = 'scalar') => {
+  const model = getPrismaModelForResource(resource);
+  const idProperty = getModelIdProperty(resource);
+
+  const displayKeys = options?.display;
+
+  let selectedFields = model?.fields.reduce(
+    (acc, field) => {
+      if ((displayKeys && displayKeys.includes(field.name as Field<M>) || !displayKeys)) {
+        if (level === 'object' && field.kind === 'object') {
+          acc[field.name] = { select: selectPayloadForModel(field.type as ModelName, {}, 'scalar') };
+        } else {
+          acc[field.name] = true;
+        }
+      }
+      return acc;
+    },
+    { [idProperty]: true } as any
+  );
+
+  return selectedFields;
+}